Commit changes in Folders, Add Special year posibility,
