+++ Queries/ISpecialTimesTollFee.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace congestion.calculator.Queries
+{
+    public interface ISpecialTimesTollFee
+    {
+        int GetFee(TimeSpan timeOfDate);
+    }
+}
