+++ Queries/IYearDayType.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace congestion.calculator.Queries
+{
+    public interface IYearDayType
+    {
+        bool IsOffDay(DateTime date);
+    }
+}
