+++ Queries/InfoOfDayToll.cs
@@ -1,4 +1,4 @@
-﻿using congestion.calculator.Services;
+﻿using congestion.calculator.Contracts;
 using System;
 
 namespace congestion.calculator.Queries
@@ -6,17 +6,28 @@ namespace congestion.calculator.Queries
     public class InfoOfDayToll
     {
         public int MaxTollFeeOfEveryDay { get; private set; }
-
+        public int SpecificYear { get; private set; }
+        public bool DoesCityHaveAnySpecificYear { get; set; }
         private readonly IYearDayType yearDayType;
         private readonly ISpecialTimesTollFee specialTimesTollFee;
 
-        public InfoOfDayToll(IYearDayType yearDayType, ISpecialTimesTollFee specialTimesTollFee)
+        public InfoOfDayToll(IYearDayType yearDayType, ISpecialTimesTollFee specialTimesTollFee, bool doesCityHaveAnySpecificYear)
         {
             this.yearDayType = yearDayType;
             this.specialTimesTollFee = specialTimesTollFee;
+            DoesCityHaveAnySpecificYear = doesCityHaveAnySpecificYear;
             GetMaxTollFeeOfEveryDay();
-
+            SetSpecificYear(doesCityHaveAnySpecificYear);
         }
+
+        private void SetSpecificYear(bool DoesCityHaveAnySpecificYear)
+        {
+            if (DoesCityHaveAnySpecificYear)
+            {
+                SpecificYear = 2013;
+            }
+        }
+
         public int GetTollFee(DateTime date, Vehicle vehicle)
         {
             if (IsTollFreeDate(date, yearDayType) || IsTollFreeVehicle(vehicle))
@@ -27,10 +38,7 @@ namespace congestion.calculator.Queries
 
         private bool IsTollFreeDate(DateTime date, IYearDayType yearDayType)
         {
-            //TODO :Clean the code
-            int year = date.Year;
-
-            if (year == 2013)
+            if (DoesCityHaveAnySpecificYear && date.Year == SpecificYear)
             {
                 return yearDayType.IsOffDay(date);
             }
