--- Services/ISpecialTimesTollFee.cs
@@ -1,6 +1,0 @@
-﻿using System;
-
-public interface ISpecialTimesTollFee
-{
-    int GetFee(TimeSpan timeOfDate);
-}
