--- Services/IYearDayType.cs
@@ -1,6 +1,0 @@
-﻿using System;
-
-public interface IYearDayType
-{ 
-    bool IsOffDay(DateTime date);
-}
